Add files via upload
Now that I've verified these work in Meat Space, enjoy!   This PCB has pinouts of each battery.  It's designed to fit on two batteries sideways and mount into the M6 on the battery.  Daly BMS or clone will fit in the M3 mounting holees in a reverse direction.

There's some minor tweaks that need to happen around routing and the drill holes, but nothing that can't be mitigated in post.